Make eval rewards independent of local environment
Rescale rewards at each time step using the terminated variable, which is normally a boolean but in the case of constraints as terminations the max constraint violation probability across all constraints. Additionally, set the hard constraint limits in eval to the values used during training. This should give an interpretable cumulative raw reward for any environment, regardless of what the local env configuration is.

Previously, if a policy was trained with higher hard constraint limits and then evaluated in the eval script with a different local env, it would return very low probabilities since the constraint violation probability was 1.0 on a lot of time steps, setting reward to 0