Ensure determinism during training on flat terrain
This restructures seeding to use a hard-coded value in the environment config, which is then extracted in train.py. After multiple days of troubleshooting, it turned out that hydra configured the environment (including running all the post_init) before running the train.py main function, which resulted in self.seed being None and thus causing nondeterministic training results. This was not noticed earlier because a) it did not occur when using a single environment and b) the changes were very small until later in the training.

The hard-coded seed is then used for seeding the libraries in train.py as well for increased transparency, even though it might be redundant. Passing via hydra at runtime did not work and would also not be a clean solution because relying on runtime args for determinism is bad practice.

Torch and physX are also adjusted to focus on determining, and a CUBLAS envvar is set to allocate memory for deterministic linear algebra. This might be overkill, but during testing the same commit produced wildly different policies so these changes should alleviate this issue and were tested on multiple cluster nodes to produce identical results after consecutive runs.

HOWEVER, this still only results in identical runs only when using flat terrain, not rough terrain, likely due to order dependent contact resolution in the PhysX engine.