Receive elevation map from python via ZMQ
Due to time constraints, need to convert elevation map topic into raw floats and send via ZMQ TCP publisher, because controller was designed to not depend on ROS2. TODO is definitely to simplify this and simply receive the topic, but this also requires integrating the elevation_mapping_cupy workspace into the main workspace and that is very problematic due to humble/jazzy compat and especially the CUDA + numpy <2.0 dependency! Life is tough